requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
frozenlist==1.7.0
genai-protocol==1.0.9
google-auth==2.40.3
google-genai==1.23.0
googletrans==4.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.6.2
openai==1.93.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
